+++ precious-core/src/filter.rs
@@ -1,6 +1,6 @@
-use crate::command;
 use crate::path_matcher;
 use anyhow::Result;
+use command;
 use log::{debug, info};
 use serde::Deserialize;
 use std::{
@@ -580,9 +580,10 @@ fn replace_root(cmd: Vec<String>, root: &Path) -> Vec<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{path_matcher, testhelper};
+    use crate::path_matcher;
     use anyhow::Result;
     use pretty_assertions::assert_eq;
+    use testhelper;
 
     type Mock = i8;